refactor: split helpers and config from app
